добавляет функцию отрисовки миниатюр
